+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.2] - 2024-06-XX
+## [0.2.2] - 2024-06-03
 ### Improved
 - Scripts will auto-escalate to sudo now. So README no longer recommends to run with sudo.
 - Logs all go to a single file now for easier viewing.
+- Reorganized directory structure.
+- Stop is now an argument './archiver.sh --stop'.
 
 ## [0.2.1] - 2024-06-02
 ### Improved